product/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
pymongo>=4.3.0
motor>=3.1.0
python-dotenv>=0.21.0